# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: string;
  name: string;
  role?: string;
  reportingMorning?: string;
  reportingAfternoon?: string;
}

export interface AttendanceLogEntry {
  timestamp: string;
  employeeName: string;
  employeeId: string;
  session: string;
  action: string;
  status: string;
  message: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
}

export interface AdminLogsResponse {
  logs: AttendanceLogEntry[];
  employees: Employee[];
}

export type LogAttendanceBodyAction =
  (typeof LogAttendanceBodyAction)[keyof typeof LogAttendanceBodyAction];

export const LogAttendanceBodyAction = {
  checkin: "checkin",
  checkout: "checkout",
} as const;

export type LogAttendanceBodySession =
  (typeof LogAttendanceBodySession)[keyof typeof LogAttendanceBodySession];

export const LogAttendanceBodySession = {
  morning: "morning",
  afternoon_out: "afternoon_out",
  afternoon_in: "afternoon_in",
  evening: "evening",
} as const;

export interface LogAttendanceBody {
  employeeId: string;
  pin: string;
  action: LogAttendanceBodyAction;
  session: LogAttendanceBodySession;
}

export type LogAttendanceResponseStatus =
  (typeof LogAttendanceResponseStatus)[keyof typeof LogAttendanceResponseStatus];

export const LogAttendanceResponseStatus = {
  on_time: "on_time",
  late: "late",
  very_late: "very_late",
  early: "early",
  on_schedule: "on_schedule",
} as const;

export interface LogAttendanceResponse {
  success: boolean;
  message: string;
  status: LogAttendanceResponseStatus;
  employeeName: string;
  timestamp: string;
}

export interface ErrorResponse {
  error: string;
}