# packages/api-schema/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all employees
 */
export const ListEmployeesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  role: zod.string().optional(),
  reportingMorning: zod.string().optional(),
  reportingAfternoon: zod.string().optional(),
});
export const ListEmployeesResponse = zod.array(ListEmployeesResponseItem);

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get all attendance logs
 */
export const GetAdminLogsHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const GetAdminLogsResponse = zod.object({
  logs: zod.array(
    zod.object({
      timestamp: zod.string(),
      employeeName: zod.string(),
      employeeId: zod.string(),
      session: zod.string(),
      action: zod.string(),
      status: zod.string(),
      message: zod.string(),
    }),
  ),
  employees: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      role: zod.string().optional(),
      reportingMorning: zod.string().optional(),
      reportingAfternoon: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Log a check-in or check-out event
 */
export const LogAttendanceBody = zod.object({
  employeeId: zod.string(),
  pin: zod.string(),
  action: zod.enum(["checkin", "checkout"]),
  session: zod.enum(["morning", "afternoon_out", "afternoon_in", "evening"]),
});

export const LogAttendanceResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  status: zod.enum(["on_time", "late", "very_late", "early", "on_schedule"]),
  employeeName: zod.string(),
  timestamp: zod.string(),
});