# packages/api-schema/src/generated/types/adminLogsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceLogEntry } from "./attendanceLogEntry";
import type { Employee } from "./employee";

export interface AdminLogsResponse {
  logs: AttendanceLogEntry[];
  employees: Employee[];
}