# packages/api-schema/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: string;
  name: string;
  role?: string;
  reportingMorning?: string;
  reportingAfternoon?: string;
}