# packages/api-schema/src/generated/types/logAttendanceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogAttendanceBodyAction } from "./logAttendanceBodyAction";
import type { LogAttendanceBodySession } from "./logAttendanceBodySession";

export interface LogAttendanceBody {
  employeeId: string;
  pin: string;
  action: LogAttendanceBodyAction;
  session: LogAttendanceBodySession;
}