# packages/api-schema/src/generated/types/logAttendanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogAttendanceResponseStatus } from "./logAttendanceResponseStatus";

export interface LogAttendanceResponse {
  success: boolean;
  message: string;
  status: LogAttendanceResponseStatus;
  employeeName: string;
  timestamp: string;
}